+++ src/core/cpu/thumb/mod.rs
@@ -6,7 +6,7 @@ pub mod thumb_impl;
 use super::ArmCpu;
 use super::super::memory::GbaMemory;
 
-pub fn execute_thumb(cpu: &mut ArmCpu, instr: u16) {
+pub fn execute_thumb(cpu: &mut ArmCpu, instr: u32) {
 	unimplemented!();
 	// let instr_sz = instr as usize;
 	// let decoded = table::THUMB_OPCODE_TABLE[(instr_sz >> 12) & 0xf][(instr_sz >> 8) & 0xf];
