+++ src/core/cpu/thumb/thumb_dp_impl.rs
@@ -3,97 +3,97 @@ use super::super::super::memory::GbaMemory;
 
 /// AND 
 /// Logical And
-pub fn thumb_dp_and(cpu: &mut ArmCpu, instr: u16) {
+pub fn thumb_dp_and(cpu: &mut ArmCpu, instr: u32) {
 	// #TODO
 }
 
 /// EOR 
 /// Logical Exclusive-or
-pub fn thumb_dp_eor(cpu: &mut ArmCpu, instr: u16) {
+pub fn thumb_dp_eor(cpu: &mut ArmCpu, instr: u32) {
 	// #TODO
 }
 
 /// LSL 
 /// Logical Left-shift
-pub fn thumb_dp_lsl(cpu: &mut ArmCpu, instr: u16) {
+pub fn thumb_dp_lsl(cpu: &mut ArmCpu, instr: u32) {
 	// #TODO
 }
 
 /// LSR 
 /// Logical Right-shift
-pub fn thumb_dp_lsr(cpu: &mut ArmCpu, instr: u16) {
+pub fn thumb_dp_lsr(cpu: &mut ArmCpu, instr: u32) {
 	// #TODO
 }
 
 /// ASR 
 /// Arithmetic Right-shift
-pub fn thumb_dp_asr(cpu: &mut ArmCpu, instr: u16) {
+pub fn thumb_dp_asr(cpu: &mut ArmCpu, instr: u32) {
 	// #TODO
 }
 
 /// ADD 
 /// Add
-pub fn thumb_dp_add(cpu: &mut ArmCpu, instr: u16) {
+pub fn thumb_dp_add(cpu: &mut ArmCpu, instr: u32) {
 	// #TODO
 }
 
 /// SUB 
 /// Subtract
-pub fn thumb_dp_sub(cpu: &mut ArmCpu, instr: u16) {
+pub fn thumb_dp_sub(cpu: &mut ArmCpu, instr: u32) {
 	// #TODO
 }
 
 /// ROR 
 /// Rotate right
-pub fn thumb_dp_ror(cpu: &mut ArmCpu, instr: u16) {
+pub fn thumb_dp_ror(cpu: &mut ArmCpu, instr: u32) {
 	// #TODO
 }
 
 /// TST 
 /// Test Bits (Logical And)
-pub fn thumb_dp_tst(cpu: &mut ArmCpu, instr: u16) {
+pub fn thumb_dp_tst(cpu: &mut ArmCpu, instr: u32) {
 	// #TODO
 }
 
 /// NEG 
 /// Negate (Subtract from zero)
-pub fn thumb_dp_neg(cpu: &mut ArmCpu, instr: u16) {
+pub fn thumb_dp_neg(cpu: &mut ArmCpu, instr: u32) {
 	// #TODO
 }
 
 /// CMP 
 /// Compare (Subtract)
-pub fn thumb_dp_cmp(cpu: &mut ArmCpu, instr: u16) {
+pub fn thumb_dp_cmp(cpu: &mut ArmCpu, instr: u32) {
 	// #TODO
 }
 
 /// CMN 
 /// Compare negative (Add)
-pub fn thumb_dp_cmn(cpu: &mut ArmCpu, instr: u16) {
+pub fn thumb_dp_cmn(cpu: &mut ArmCpu, instr: u32) {
 	// #TODO
 }
 
 /// ORR 
 /// Logical Or
-pub fn thumb_dp_orr(cpu: &mut ArmCpu, instr: u16) {
+pub fn thumb_dp_orr(cpu: &mut ArmCpu, instr: u32) {
 	// #TODO
 }
 
 /// MUL 
 /// Multiply
-pub fn thumb_dp_mul(cpu: &mut ArmCpu, instr: u16) {
+pub fn thumb_dp_mul(cpu: &mut ArmCpu, instr: u32) {
 	// #TODO
 }
 
 /// BIC 
 /// Bit Clear (NAND)
-pub fn thumb_dp_bic(cpu: &mut ArmCpu, instr: u16) {
+pub fn thumb_dp_bic(cpu: &mut ArmCpu, instr: u32) {
 	// #TODO
 }
 
 /// MVN 
 /// Move negative (NOT)
-pub fn thumb_dp_mvn(cpu: &mut ArmCpu, instr: u16) {
+pub fn thumb_dp_mvn(cpu: &mut ArmCpu, instr: u32) {
 	// #TODO
 }
